Change Ordinal from int to i64

+++ core/time/datetime/internal.odin
@@ -4,7 +4,7 @@ package datetime
 
 import "base:intrinsics"
 
-sign :: proc "contextless" (v: int) -> (res: int) {
+sign :: proc "contextless" (v: i64) -> (res: i64) {
 	if v == 0 {
 		return 0
 	} else if v > 0 {
@@ -37,7 +37,7 @@ floor_div :: proc "contextless" (x, y: $T) -> (res: T)
 }
 
 // Half open: x mod [1..b]
-interval_mod :: proc "contextless" (x, a, b: int) -> (res: int) {
+interval_mod :: proc "contextless" (x, a, b: i64) -> (res: i64) {
 	if a == b {
 		return x
 	}
@@ -45,12 +45,12 @@ interval_mod :: proc "contextless" (x, a, b: int) -> (res: int) {
 }
 
 // x mod [1..b]
-adjusted_remainder :: proc "contextless" (x, b: int) -> (res: int) {
+adjusted_remainder :: proc "contextless" (x, b: i64) -> (res: i64) {
 	m := x %% b
 	return b if m == 0 else m
 }
 
-gcd :: proc "contextless" (x, y: int) -> (res: int) {
+gcd :: proc "contextless" (x, y: i64) -> (res: i64) {
 	if y == 0 {
 		return x
 	}
@@ -59,18 +59,18 @@ gcd :: proc "contextless" (x, y: int) -> (res: int) {
 	return gcd(y, m)
 }
 
-lcm :: proc "contextless" (x, y: int) -> (res: int) {
+lcm :: proc "contextless" (x, y: i64) -> (res: i64) {
 	return x * y / gcd(x, y)
 }
 
-sum :: proc "contextless" (i: int, f: proc "contextless" (n: int) -> int, cond: proc "contextless" (n: int) -> bool) -> (res: int) {
+sum :: proc "contextless" (i: i64, f: proc "contextless" (n: i64) -> i64, cond: proc "contextless" (n: i64) -> bool) -> (res: i64) {
 	for idx := i; cond(idx); idx += 1 {
 		res += f(idx)
 	}
 	return
 }
 
-product :: proc "contextless" (i: int, f: proc "contextless" (n: int) -> int, cond: proc "contextless" (n: int) -> bool) -> (res: int) {
+product :: proc "contextless" (i: i64, f: proc "contextless" (n: i64) -> i64, cond: proc "contextless" (n: i64) -> bool) -> (res: i64) {
 	res = 1
 	for idx := i; cond(idx); idx += 1 {
 		res *= f(idx)
@@ -78,7 +78,7 @@ product :: proc "contextless" (i: int, f: proc "contextless" (n: int) -> int, co
 	return
 }
 
-smallest :: proc "contextless" (k: int, cond: proc "contextless" (n: int) -> bool) -> (d: int) {
+smallest :: proc "contextless" (k: i64, cond: proc "contextless" (n: i64) -> bool) -> (d: i64) {
 	k := k
 	for !cond(k) {
 		k += 1
@@ -86,7 +86,7 @@ smallest :: proc "contextless" (k: int, cond: proc "contextless" (n: int) -> boo
 	return k
 }
 
-biggest :: proc "contextless" (k: int, cond: proc "contextless" (n: int) -> bool) -> (d: int) {
+biggest :: proc "contextless" (k: i64, cond: proc "contextless" (n: i64) -> bool) -> (d: i64) {
 	k := k
 	for !cond(k) {
 		k -= 1
